Move orchestrator code out of library
