+++ hirte/src/orch/main.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "../../../libhirte/src/orchestrator.h"
+#include "orchestrator.h"
 #include "../../../libhirte/src/service/shutdown.h"
 #include "../ini/config.h"
 #include "opt.h"
